requirements.txt
streamlit==1.28.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pandas==2.1.3
plotly==5.18.0
bcrypt==4.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
